Detect PHP version in IDE
